web_projects/django_projects/todos/requirements.txt
django==4.1.6
django-crispy-forms==1.14.0
crispy-bootstrap5
# komentarz
# -r core.txt
django-import-export
pillow
faker
sorl-thumbnail
python-memcached
